requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
pydantic[email]>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.7
